backend/pyproject.toml
[project]
name = "apply-srbiau"
version = "0.1.0"
description = "Student application journey sharing platform - Learn from those who applied before you"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "python-multipart>=0.0.9",
    "aiofiles>=24.1.0",
    "pydantic-settings>=2.4.0",
    "python-jose[cryptography]>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
